requirements.txt
asgiref==3.4.1
autopep8==1.5.7
blosc2==2.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==3.2.6
django-import-export==4.1.1
django-tables2==2.7.0
django-widget-tweaks==1.5.0
gunicorn==20.1.0
msgpack==1.1.0
ndindex==1.8
numexpr==2.10.1
numpy==2.1.1
packaging==24.1
py-cpuinfo==9.0.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
tables==3.10.1
tablib==3.5.0
toml==0.10.2
typing_extensions==4.12.2
Unipath==1.1
whitenoise==5.3.0
